sstables: add mirrored_storage
this change adds yet another storage backend to SSTables subsystem.
the storage backend is named "mirrored_storage", because it writes
to both the local filesystem and to the remote object storage.
this new storage backend will be used as a solution of the backup,
so for instance we can replace a scylla instance from the sstables
stored in object storage with a new node. and in future, based on
this storage backend we can add more compolicated policy, for instance,
to optionally store some sstables / some components / some ranges of
data  locally, for better cost, for better availability by hiding
the complexity in the storage, and by exposing the consistent
i/o interface to the users of sstables.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>