test: test_topology_ops: disable background writes
Recently, in a3ba4b31098074a24ce1b3c941341a5375dbbe14, this test was
extended with a background task that continuously performs CQL writes.

This turned out to be very valuable and detected a couple of bugs,
including:
https://github.com/scylladb/scylladb/issues/15924
https://github.com/scylladb/scylladb/issues/15935

Unfortunately this causes CI to be flaky.
Until these bugs are fixed, we disable the background writes to unflake
CI.

Closes scylladb/scylladb#15937