raft topology: handle abort exceptions from sleeping in fence_previous_coordinator
The fence_previous_coordinator function has a retry loop: if it fails to
perform a group0 operation, it will try again after a 1 second delay.
However, if the topology coordinator is aborted while it waits, an
exception will be thrown and will be propagated out of the function. The
function is supposed to handle all exceptions internally, so this is not
desired.

Fix this by catching the abort_requested_exception and returning from
the function if the exception is caught.